requirements/common.txt
Django
django-cors-headers
djangorestframework
drf-yasg
Pillow
python-dotenv
python-jose
requests

